refactor: change directory and delete useless files
