+++ .gitignore
@@ -1,0 +1,6 @@
+# 다음과 같은 확장자는 무시
+
+*.csv
+*.pkl
+*.docx
+*.pdf
