--- heuristic_movie_genre_scores.csv
@@ -1,31 +1,0 @@
-,genre,score
-0,드라마,"[0, 0, 0, 0, 1, -1, 1, -1]"
-1,판타지,"[0, 0, -1, 1, 1, -1, 1, -1]"
-2,서부,"[1, -1, 1, -1, 1, -1, 1, -1]"
-3,공포,"[-1, 1, 1, -1, 0, 0, 0, 0]"
-4,멜로,"[-1, 1, 0, 0, 1, -1, 0, 0]"
-5,로맨스,"[1, -1, 0, 0, 1, -1, 0, 0]"
-6,모험,"[1, -1, -1, 1, 0, 0, 1, -1]"
-7,스릴러,"[1, -1, 1, -1, 0, 0, 0, 0]"
-8,느와르,"[0, 0, 1, -1, 0, 0, 0, 0]"
-9,컬트,"[-1, 1, 0, 0, -1, 1, 0, 0]"
-10,다큐멘터리,"[-1, 1, 1, -1, -1, 1, -1, 1]"
-11,코미디,"[1, -1, 0, 0, 1, -1, 0, 0]"
-12,가족,"[0, 0, 0, 0, 1, -1, 0, 0]"
-13,미스터리,"[-1, 1, 1, -1, -1, 1, 0, 0]"
-14,전쟁,"[0, 0, 1, -1, -1, 1, -1, 1]"
-15,애니메이션,"[0, 0, 0, 0, 1, -1, 0, 0]"
-16,범죄,"[0, 0, 1, -1, -1, 1, 0, 0]"
-17,뮤지컬,"[1, -1, -1, 1, 1, -1, 1, -1]"
-18,SF,"[0, 0, -1, 1, -1, 1, 0, 0]"
-19,액션,"[1, -1, 0, 0, 1, -1, 0, 0]"
-20,무협,"[0, 0, 0, 0, 1, -1, 0, 0]"
-21,에로,"[-1, 1, 1, -1, 1, -1, 0, 0]"
-22,서스펜스,"[-1, 1, 1, -1, 0, 0, 0, 0]"
-23,서사,"[-1, 1, 1, -1, 1, -1, 0, 0]"
-24,블랙코미디,"[1, -1, 1, -1, 0, 0, 0, 0]"
-25,실험,"[0, 0, -1, 1, 0, 0, 1, -1]"
-26,영화카툰,"[1, -1, 0, 0, 1, -1, 1, -1]"
-27,영화음악,"[-1, 1, 1, -1, 1, -1, 0, 0]"
-28,영화패러디포스터,"[0, 0, -1, 1, 1, -1, 0, 0]"
-29,공연실황,"[1, -1, 1, -1, 1, -1, 0, 0]"
