--- heuristic_music_theme_scores.csv
@@ -1,73 +1,0 @@
-,theme,E,I,S,N,F,T,P,J
-0,신나는,1,-1,0,0,1,-1,1,-1
-1,강한,1,-1,0,0,0,0,0,0
-2,감성적인,-1,1,-1,1,1,-1,1,-1
-3,공연/라이브,0,0,-1,1,1,-1,0,0
-4,봄,0,0,1,-1,1,-1,0,0
-5,걸그룹,1,-1,0,0,0,0,0,0
-6,소울풀한,-1,1,0,0,1,-1,0,0
-7,여름,1,-1,1,-1,1,-1,0,0
-8,듀엣/피처링,0,0,0,0,1,-1,0,0
-9,아침,0,0,0,0,1,-1,0,0
-10,겨울,-1,1,1,-1,1,-1,0,0
-11,아이돌,1,-1,0,0,0,0,0,0
-12,가을,-1,1,1,-1,1,-1,0,0
-13,목소리/음색,-1,1,0,0,0,0,0,0
-14,청량한,1,-1,1,-1,0,0,0,0
-15,잔잔한,-1,1,0,0,1,-1,0,0
-16,올디스,0,0,1,-1,1,-1,-1,1
-17,섹시한,0,0,-1,1,1,-1,1,-1
-18,애절한,-1,1,1,-1,1,-1,0,0
-19,감각적인,-1,1,1,-1,0,0,0,0
-20,쓸쓸한,-1,1,1,-1,1,-1,0,0
-21,달달한,0,0,1,-1,1,-1,0,0
-22,그루브한,0,0,-1,1,0,0,1,-1
-23,어쿠스틱한,-1,1,-1,1,1,-1,0,0
-24,발라드한,-1,1,0,0,1,-1,0,0
-25,일렉트로닉,1,-1,-1,1,-1,1,0,0
-26,몽환적인,-1,1,-1,1,-1,1,0,0
-27,오후,0,0,0,0,1,-1,0,0
-28,저녁,0,0,0,0,1,-1,0,0
-29,신선한,1,-1,1,-1,0,0,0,0
-30,미세먼지/황사,1,-1,1,-1,0,0,0,0
-31,하교/퇴근길,1,-1,-1,1,0,0,1,-1
-32,휴식/명상,-1,1,-1,1,-1,1,-1,1
-33,드라이브,1,-1,0,0,1,-1,0,0
-34,쌀쌀한 날,-1,1,1,-1,0,0,0,0
-35,운동/헬스,1,-1,-1,1,-1,1,-1,1
-36,폭염/더위,1,-1,-1,1,0,0,0,0
-37,크리스마스,1,-1,1,-1,1,-1,0,0
-38,비온후/맑게갠,-1,1,-1,1,1,-1,1,-1
-39,밤/새벽,-1,1,-1,1,1,-1,1,-1
-40,화창한 날,1,-1,1,-1,1,-1,0,0
-41,환절기,-1,1,0,0,1,-1,1,-1
-42,눈오는 날,-1,1,1,-1,1,-1,0,0
-43,비/흐림,-1,1,-1,1,1,-1,1,-1
-44,노래방,1,-1,-1,1,0,0,1,-1
-45,편집숍/매장,-1,1,-1,1,1,-1,1,-1
-46,호텔/바,-1,1,-1,1,-1,1,1,-1
-47,방안에서,-1,1,-1,1,1,-1,1,-1
-48,잠들기 전,-1,1,0,0,1,-1,1,-1
-49,카페,-1,1,-1,1,1,-1,1,-1
-50,산책/여행,1,-1,1,-1,1,-1,1,-1
-51,독서,-1,1,-1,1,-1,1,-1,1
-52,결혼,-1,1,1,-1,1,-1,0,0
-53,클럽/파티,1,-1,-1,1,0,0,1,-1
-54,페스티벌,1,-1,0,0,0,0,1,-1
-55,패션쇼,0,0,-1,1,-1,1,0,0
-56,혼술혼밥,-1,1,-1,1,-1,1,0,0
-57,그리움,-1,1,0,0,1,-1,0,0
-58,사랑/기쁨,-1,1,-1,1,1,-1,0,0
-59,에스테틱,-1,1,-1,1,-1,1,0,0
-60,외로울때,-1,1,0,0,1,-1,0,0
-61,새벽감성,-1,1,-1,1,1,-1,1,-1
-62,썸 탈때,0,0,0,0,1,-1,1,-1
-63,설렘/심쿵,0,0,0,0,1,-1,1,-1
-64,기분전환,1,-1,0,0,1,-1,1,-1
-65,싱숭생숭,-1,1,-1,1,1,-1,1,-1
-66,울고 싶을때,-1,1,1,-1,1,-1,0,0
-67,고백,1,-1,0,0,1,-1,1,-1
-68,멘붕/불안,-1,1,0,0,-1,1,1,-1
-69,지치고 힘들때,-1,1,-1,1,1,-1,0,0
-70,우울할때,-1,1,0,0,1,-1,0,0
-71,스트레스/짜증,1,-1,-1,1,0,0,1,-1
