+++ score/한국어불용어.txt
@@ -1,0 +1,155 @@
+형태	품사	비율
+이	VCP	0.018279601
+있	VA	0.011699048
+하	VV	0.009773658
+것	NNB	0.00973315
+들	XSN	0.00689824
+그	MM	0.005327252
+되	VV	0.00361335
+수	NNB	0.003473622
+이	NP	0.003361203
+보	VX	0.003310379
+않	VX	0.0029757
+없	VA	0.002920476
+나	NP	0.002690479
+사람	NNG	0.002073917
+주	VV	0.001884579
+아니	VCN	0.001870774
+등	NNB	0.001822074
+같	VA	0.001724522
+우리	NP	0.001714509
+때	NNG	0.001685684
+년	NNB	0.001647755
+가	VV	0.001618626
+한	MM	0.001584036
+지	VX	0.001537763
+대하	VV	0.00150378
+오	VV	0.001490884
+말	NNG	0.001321876
+일	NNG	0.001239799
+그렇	VA	0.001232062
+위하	VV	0.001170618
+때문	NNB	0.001146647
+그것	NP	0.001099161
+두	VV	0.001063053
+말하	VV	0.001030587
+알	VV	0.000995693
+그러나	MAJ	0.000978246
+받	VV	0.000924084
+못하	VX	0.000904665
+일	NNB	0.000880239
+그런	MM	0.000879481
+또	MAG	0.000870985
+문제	NNG	0.000823954
+더	MAG	0.000801652
+사회	NNG	0.000793459
+많	VA	0.000789211
+그리고	MAJ	0.000779653
+좋	VA	0.000763723
+크	VA	0.000758565
+따르	VV	0.000738084
+중	NNB	0.000733229
+나오	VV	0.000725188
+가지	VV	0.000720485
+씨	NNB	0.000710169
+시키	XSV	0.000707893
+만들	VV	0.000703797
+지금	NNG	0.000699701
+생각하	VV	0.000694846
+그러	VV	0.000692418
+속	NNG	0.00068544
+하나	NR	0.000655097
+집	NNG	0.000648573
+살	VV	0.000648422
+모르	VV	0.000645084
+적	XSN	0.000643112
+월	NNB	0.000639319
+데	NNB	0.000636436
+자신	NNG	0.000632188
+안	MAG	0.000625968
+어떤	MM	0.000623237
+내	NP	0.000622023
+내	VV	0.000614438
+경우	NNG	0.000613831
+명	NNB	0.00061019
+생각	NNG	0.000587433
+시간	NNG	0.00057302
+그녀	NP	0.000571806
+다시	MAG	0.00055618
+이런	MM	0.000550263
+앞	NNG	0.000548898
+보이	VV	0.000542678
+번	NNB	0.000542526
+나	VX	0.000538126
+다른	MM	0.000536306
+어떻	VA	0.000533423
+여자	NNG	0.000531754
+개	NNB	0.000530085
+전	NNG	0.00052872
+들	VV	0.000514914
+사실	NNG	0.000511728
+이렇	VA	0.000511576
+점	NNG	0.000507935
+싶	VX	0.000504749
+말	VX	0.000503991
+정도	NNG	0.000496102
+좀	MAG	0.000492916
+원	NNB	0.000492309
+잘	MAG	0.000491399
+통하	VV	0.000486847
+소리	NNG	0.000485633
+놓	VX	0.000482751
+에
+의
+니다
+다
+도
+을
+보다
+햐
+였다
+않고
+이었습니다
+됬네요
+와
+는
+너무
+인
+로
+있는
+하고
+했는데
+스러우며
+가는줄
+있군요
+요
+를
+으로
+며
+과
+할
+분
+과
+하는
+에도
+도
+고
+입니다
+봤어요
+며
+라
+이렇게
+네
+시
+만
+된
+뭐
+뿐
+서서
+ㄴ
+었
+에게
+게
+임
+은
+는
